Задание 2 На коррекцию

+++ src/main/java/Main/java/Category.java
@@ -10,7 +10,7 @@ public class Category {
     @Column(nullable = false)
     private String name;/*- название категории*/
     @Column
-    private Integer parent;/*- родительская категория*/
+    private Long parent;/*- родительская категория*/
 
     public Long getId() {
         return id;
@@ -28,11 +28,11 @@ public class Category {
         this.name = name;
     }
 
-    public Integer getParent() {
+    public Long getParent() {
         return parent;
     }
 
-    public void setParent(Integer parent) {
+    public void setParent(Long parent) {
         this.parent = parent;
     }
 }
